pyproject.toml
[project]
name = "klabban"
version = "0.1.0"
description = ""
authors = [
    {name = "Nopphakao Panjasataros",email = "124285253+krahyor@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "flask-mongoengine-3 (>=1.1.0,<2.0.0)",
    "flask-login (>=0.6.3,<0.7.0)",
    "flask-bcrypt (>=1.0.1,<2.0.0)",
    "flask-wtf (>=1.2.2,<2.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "uuid6 (>=2025.0.1,<2026.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "xlsxwriter (>=3.2.9,<4.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "rq (>=2.6.0,<3.0.0)",
    "user-agents (>=2.2.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
klabban-web = "klabban.cmd.web:main"
klabban-worker = "klabban.cmd.worker:main"